+++ app/persistence/models/submission_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Text, DateTime, ForeignKey, UniqueConstraint
+from sqlalchemy import Column, Integer, Text, DateTime, ForeignKey, UniqueConstraint, BigInteger
 from sqlalchemy.sql import func
 from app.persistence.database import Base
 
@@ -7,7 +7,7 @@ class Submission(Base):
 
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     proposal_id = Column(Integer, ForeignKey("proposals.id"), nullable=False, index=True)
-    submitter_id = Column(Integer, ForeignKey("users.telegram_id"), nullable=False, index=True)
+    submitter_id = Column(BigInteger, ForeignKey("users.telegram_id"), nullable=False, index=True)
     response_content = Column(Text, nullable=False)
     timestamp = Column(DateTime(timezone=True), server_default=func.now(), nullable=False)
